requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
db-dtypes==1.5.0
google-api-core==2.30.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
numpy==2.4.3
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pandas-gbq==0.34.0
patsy==1.0.2
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
requests-oauthlib==2.0.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
statsmodels==0.14.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
